fix code style
